Move exceptions to their own header
- Avoid circular dependencies
- Use SolverException in Matrix class
